VAN-10860 fixed opened gaming subpage that opened with protocol show dashboard page

+++ vtr-ui/src/app/components/pages/page-device-settings/page-device-settings.component.ts
@@ -282,8 +282,12 @@ export class PageDeviceSettingsComponent implements OnInit, OnDestroy {
 		if (this.notificationSubscription) {
 			this.notificationSubscription.unsubscribe();
 		}
-		this.qaService.destroyChangeSubscribed();
-		this.routerSubscription.unsubscribe();
+		if (this.qaService) {
+			this.qaService.destroyChangeSubscribed();
+		}
+		if (this.routerSubscription) {
+			this.routerSubscription.unsubscribe();
+		}
 	}
 
 }
